==> wrangling/2-weather-wrangling.Rmd <==
---
title: "Weather Data Wrangling"
author: "Eric R. Scott"
date: "2019-11-5"
output: 
  html_notebook: 
    highlight: kate
    theme: yeti
    toc: yes
    toc_float: yes
    number_sections: yes
---
```{r setup, message=FALSE, warning=FALSE}
library(tidyverse)
library(readxl)
library(here)
library(lubridate)
library(glue)
library(conflicted)
library(tsibble)
library(tsModel)

conflict_prefer("here", "here")
conflict_prefer("filter", "dplyr")
conflict_prefer("lag", "dplyr")
conflict_prefer("interval", "lubridate")
```

# Purpose

There are essentially two sources of data I will use in this document.  1) the data taken *on site* at Shanfu Tea Company, by me, and 2) local weather station data.  The on-site data spans most of June and July and is hourly temperature data from a HOBO logger and daily precipitation data (from only mid-June onwards) from a rain gauge.  The local weather station data is transcribed from hand-written data sheets sent to me as images by Dr. Han and extends backwards in time to May.  I will wrangle both data sources and compare them.

To fill in a couple of NAs, I'll use a third data source---county-level data provided by Dr. Han.  He said this is less accurate than the local weather station data, but it will be useful for a couple of missing values in the local precip data.

# Read in data

```{r}
hobo <- read_rds(here("data", "cleaned", "daily-temps-hobo.rds"))
rain_gauge <- read_rds(here("data", "cleaned", "rain-gauge.rds"))

caas_weather <- read_rds(here("data", "cleaned", "daily-weather-caas.rds"))

county_weather <- 
  read_excel(here("data", "Shaxian climate data in 2016-2017.xlsx"), skip = 1,
             col_types = c("date", "numeric", "numeric", "numeric", "numeric", "numeric", "numeric")) %>%
  filter(!is.na(Date)) %>% 
  mutate(date = as.Date(Date),
         precip = `rainfall (0.1 mm)`) %>% #pretty sure units are in mm and this just means they were measured out to 0.1 mm.
  filter(date > ymd("2017-05-01")) %>% 
  select(-Date, - `rainfall (0.1 mm)`)
```

# Join Data Sources

CAAS weather station data and on-site hobo data logger

```{r}
temperature <-
  full_join(caas_weather, hobo, by = "date")
```
```{r}
cor.test(temperature$temp_mean_caas, temperature$temp_mean)
temperature %>% 
  transmute(temp_diff = temp_mean_caas - temp_mean) %>%
  mutate(square_diff = temp_diff^2) %>% 
  filter(complete.cases(.)) %>% 
  summarize(RMSE = sqrt(mean(square_diff)))
```

Join rain gauge data

```{r}
weather_full <- full_join(temperature, rain_gauge, by = "date")
```

# Compare data sources

I just want to check how reliable the data sources are by plotting them together

## Temperature comparison

```{r}
ggplot(weather_full, aes(x = date)) +
  geom_line(aes(y = temp_mean_caas), color = "red") +
  geom_line(aes(y = temp_mean), color = "darkred") + 
  labs(y = "mean temp (ºC)")
```
Tempreature data from weather station matches HOBO data logger nearly perfectly.  That means I can use HOBO data for the period I have it, and prepend with weather station data.  This could be a supplemental figure.


## Precipitation comparison

I think the units for one of these sources must be wrong.  I'm confused.

```{r}
ggplot(weather_full, aes(x = date)) +
  geom_line(aes(y = precip_caas_mm), color = "blue") +
  geom_line(aes(y = precip_mm/10), color = "purple") +#spreadsheet says units of 0.1mm
  geom_line(data = county_weather, aes(x = date, y = precip), color = "darkgreen")
```

Well, these don't match up as well, but I really have no choice other than to use the weather station data since the rain gauge data is so limited.

# Combine data sources

For temperature, I'll use the HOBO data and prepend with CAAS weather station data.  For precipitation, I'll use only the CAAS weather station data.

```{r}
weather <- 
  weather_full %>% 
  #cutoff at last date of experiment
  filter(date <= ymd("2017-07-24")) %>% 
  mutate(temp_mean = ifelse(is.na(temp_mean), temp_mean_caas, temp_mean),
         temp_min = ifelse(is.na(temp_min), temp_min_caas, temp_min),
         temp_max = ifelse(is.na(temp_max), temp_max_caas, temp_max)) %>% 
  select(date, temp_mean, temp_min, temp_max, precip = precip_caas_mm)
```

## Dealing with missing values

```{r}
weather %>% filter(is.na(temp_mean) | is.na(precip))
```
```{r}
rain_gauge %>% filter(date %within% interval(ymd("2017-06-27"), ymd("2017-06-28")))
county_weather %>% filter(date %within% interval(ymd("2017-06-27"), ymd("2017-06-28")))
```


06-27 and 06-28 have NAs for precip.  It's unclear from the raw data sheet if the precip measurments were 0's or not recorded for this date.  According to the county level data, it did rain, so I'll swap those data in. (data are in units of 0.1mm, so 7.5 = 0.75mm).

```{r}
weather <- weather %>% 
  mutate(precip = ifelse(date == ymd("2017-06-27"), 0.75, precip)) %>%
  mutate(precip = ifelse(date == ymd("2017-06-28"), 0.5, precip))
```

For the two missing days of temperature data, I'll use imputed values (see appendix).

```{r}
caas_imp <- read_rds(here("data", "cleaned", "daily-weather-caas-imputed.rds"))
weather <- 
  left_join(weather, caas_imp) %>% 
  mutate(temp_mean = ifelse(is.na(temp_mean), temp_mean_caas, temp_mean),
         temp_min = ifelse(is.na(temp_min), temp_min_caas, temp_min),
         temp_max = ifelse(is.na(temp_max), temp_max_caas, temp_max)) %>% 
  select(-contains("_caas"))
```


# Create lagged variables

```{r paged.print=FALSE}
nlags = 15
weather_lags <- 
  as_tsibble(weather, index = date) %>% 
  tsibble::fill_gaps() %>% 
  mutate(Q_precip = Lag(precip, 1:nlags),
         Q_temp_mean = Lag(temp_mean, 1:nlags),
         Q_temp_min = Lag(temp_min, 1:nlags),
         Q_temp_max = Lag(temp_max, 1:nlags)) %>% 
  mutate(L = matrix(1:nlags, byrow=TRUE, ncol = nlags, nrow = n()))
```


# Write data

```{r}
write_rds(weather_lags, here("data", "cleaned", "weather_lags.rds"))
```
