==> wrangling/3-plant-wrangling.Rmd <==
---
title: "Plant growth data wrangling"
author: "Eric R. Scott"
date: "2019-11-5"
output: 
  html_notebook: 
    highlight: kate
    theme: yeti
    toc: yes
    toc_float: yes
    number_sections: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, paged.print = FALSE)
```

*Last compiled: `r Sys.Date()`*

```{r packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(lubridate)
library(conflicted)
conflict_prefer("filter", "dplyr")
conflict_prefer("lag", "dplyr")
conflict_prefer("here", "here")
```

# Purpose

gathering of shoots by plant and date, and reading in of stem diameter data done in the shoot growth script. One challenge is that every inter-harvest period the shoots are re-set.  That's why in some places I've split by field and then recombined.

# Read in data

```{r}
growth <- read_rds(here("data", "cleaned", "shoot-growth.rds"))
diameter <- read_rds(here("data", "cleaned", "shoot-diameter.rds"))
weather_lags <- read_rds(here("data", "cleaned", "weather_lags.rds"))
```

`growth`: field data including plant growth and leafhopper densities
`diameter`: stem diameters taken once per harvest
`weather_lags`: output from "2-weather-wrangling.Rmd"

Separate fields A and B

```{r, message=TRUE, warning=TRUE}
input.A <- growth %>% filter(field == "A")
input.B <- growth %>% filter(field == "B")
```

# Create inter-harvest periods

Find harvest dates for both fields

```{r find harvest dates}
#Find the harvest dates, i.e. the first date of each inter-harvest period.

harvest.dates.A <-
  input.A %>% 
  group_by(date) %>% 
  summarise(Harvest = last(harvest)) %>% 
  filter(Harvest == "Y") %>% 
  ungroup() %>% 
  select(date)

harvest.dates.B <- input.B %>% 
  group_by(date) %>% 
  summarise(Harvest = last(harvest)) %>% 
  filter(Harvest == "Y") %>% 
  ungroup() %>% 
  select(date)

harvest.dates.B <- add_row(harvest.dates.B, date = max(input.B$date) + days(1))
#on the last day of data collection field B was harvested
```

Create harvest intervals.  Intervals are a data format from the `lubridate` package.  Because of the way I've coded harvesting, the last date of each inter-harvest interval is the harvest date - 1 day.

```{r create harvest intervals}
harvest.int.A <- (lag(harvest.dates.A$date) - days(1)) %--% (harvest.dates.A$date - days(1))
harvest.int.A <- harvest.int.A[-1]

harvest.int.B <- (lag(harvest.dates.B$date) - days(1)) %--% (harvest.dates.B$date - days(1))

#the very first day of data collection was not immediately after a harvest.  Add that interval manually.

harvest.int.B[1] <- min(input.B$date) %--% (harvest.dates.B$date[1] - days(1))
```

Ok, now I create a column showing which interharvest period each datapoint belongs to labeled by harvest date and another column labeled by harvest number (1-4)

```{r assign intervals}
data.A <-
  input.A %>%
  mutate(interharvest = case_when(date %within% harvest.int.A[1] ~ int_end(harvest.int.A[1]),
                                  date %within% harvest.int.A[2] ~ int_end(harvest.int.A[2]),
                                  date %within% harvest.int.A[3] ~ int_end(harvest.int.A[3]),
                                  date %within% harvest.int.A[4] ~ int_end(harvest.int.A[4]),
                                  TRUE                           ~ as.POSIXct(NA)),
         interharvest.no = case_when(date %within% harvest.int.A[1] ~ 1,
                                     date %within% harvest.int.A[2] ~ 2,
                                     date %within% harvest.int.A[3] ~ 3,
                                     date %within% harvest.int.A[4] ~ 4,
                                     TRUE                           ~ as.numeric(NA)) %>% as.factor())

# Field A data collection ended before field B.  Now that these data frames are split, data.A has a bunch of NAs.  I'll get rid of them, although I'll need to remember this when re-joining the data.

data.A <- data.A %>% filter(date < ymd("2017-07-22"))

data.B <- input.B %>%
  mutate(interharvest = case_when(date %within% harvest.int.B[1] ~ int_end(harvest.int.B[1]),
                                  date %within% harvest.int.B[2] ~ int_end(harvest.int.B[2]),
                                  date %within% harvest.int.B[3] ~ int_end(harvest.int.B[3]),
                                  date %within% harvest.int.B[4] ~ int_end(harvest.int.B[4]),
                                  TRUE                           ~ as.POSIXct(NA)),
         interharvest.no = case_when(date %within% harvest.int.B[1] ~ 1,
                                     date %within% harvest.int.B[2] ~ 2,
                                     date %within% harvest.int.B[3] ~ 3,
                                     date %within% harvest.int.B[4] ~ 4,
                                     TRUE                           ~ as.numeric(NA)) %>% as.factor())

```


Now I group by interhavest period and calculate a growth rate for each shoot on each day.

Preliminary analysis leads me to believe a linear growth rate is best ($h_2 - h_1$), but I'll also calculate a geometric growth rate ($r = h_2/h_1$)

```{r}
data.B.full <- 
  data.B %>% 
  mutate(shoot_id = paste0("f", field, "-h", interharvest.no, "-p", plant, "-s", shoot)) %>% 
  group_by(interharvest, shoot_id) %>% 
  arrange(date) %>% 
  mutate(growth = shoot_height - lag(shoot_height),
         r = shoot_height/lag(shoot_height))

data.A.full <- 
  data.A %>% 
  mutate(shoot_id = paste0("f", field, "-h", interharvest.no, "-p", plant, "-s", shoot)) %>% 
  group_by(interharvest, shoot_id) %>% 
  arrange(date) %>% 
  mutate(growth = shoot_height - lag(shoot_height),
         r = shoot_height/lag(shoot_height))
```


# Join with stem diameter data

Now the trick is joining it to the rest of the data correctly. Since the date the stem diameter was measured on isn't the first day of measurements always, I should do the same interharvest thing I did before and join by that column.  FIRST, I have to split by field in order to do that.

```{r}
dia.data.A <-
  diameter %>% 
  filter(field == "A") %>%
  mutate(interharvest = case_when(date %within% harvest.int.A[1] ~ int_end(harvest.int.A[1]),
                                  date %within% harvest.int.A[2] ~ int_end(harvest.int.A[2]),
                                  date %within% harvest.int.A[3] ~ int_end(harvest.int.A[3]),
                                  date %within% harvest.int.A[4] ~ int_end(harvest.int.A[4]),
                                  TRUE                           ~ as.POSIXct(NA)),
         interharvest.no = case_when(date %within% harvest.int.A[1] ~ 1,
                                     date %within% harvest.int.A[2] ~ 2,
                                     date %within% harvest.int.A[3] ~ 3,
                                     date %within% harvest.int.A[4] ~ 4,
                                     TRUE                           ~ as.numeric(NA)) %>% as.character())
```


```{r}
dia.data.B <- 
  diameter %>%
  filter(field == "B") %>% 
  mutate(interharvest = case_when(date %within% harvest.int.B[1] ~ int_end(harvest.int.B[1]),
                                  date %within% harvest.int.B[2] ~ int_end(harvest.int.B[2]),
                                  date %within% harvest.int.B[3] ~ int_end(harvest.int.B[3]),
                                  date %within% harvest.int.B[4] ~ int_end(harvest.int.B[4]),
                                  TRUE                           ~ as.POSIXct(NA)),
         interharvest.no = case_when(date %within% harvest.int.B[1] ~ 1,
                                     date %within% harvest.int.B[2] ~ 2,
                                     date %within% harvest.int.B[3] ~ 3,
                                     date %within% harvest.int.B[4] ~ 4,
                                     TRUE                           ~ as.numeric(NA)) %>% as.character())

```


```{r}
#merge it back together and overwrite
dia.data <- bind_rows(dia.data.A, dia.data.B)
#so actually the date I measured stem diameter doesn't matter
dia.data2 <-
  dia.data %>% 
  mutate(shoot_id = paste0("f", field, "-h", interharvest.no, "-p", plant, "-s", shoot)) %>% 
  select(shoot_id, diameter)
```

Now I can join it with the shoot data

```{r}
shoot.data <- bind_rows(data.A.full, data.B.full) %>% left_join(dia.data2)
```

# Add weather data

```{r}
shoot_data_full <-
  left_join(shoot.data, weather_lags, by = "date") %>% 
  arrange(date)
```


## Pre-treatment
1. Create variable for # days post harvest and variable for start date.

```{r}
shoot.data2 <-
  shoot_data_full %>%
  group_by(shoot_id) %>% 
  mutate(day_post = day - min(day, na.rm = TRUE),
         start = min(day, na.rm = TRUE),
         mean_day = mean(day, na.rm = TRUE)) %>%
  ungroup()
```

3. create unique IDs for plant and interharvest number nested within field

```{r}
shoot.data3 <-
  shoot.data2 %>% 
  mutate(plant_id = paste0(field, plant),
         interharvest_id = paste0(field, interharvest.no)) 
```

4. Remove first interharvest period for field B.  It was only 4 days, so I think data collection started in the middle of an interharvest period, making `day_post` unreliable as a co-variate for this replicate.

```{r}
shoot.data4 <- shoot.data3 %>% filter(interharvest_id != "B1")
```


5. Remove outliers

```{r}
shoot.data5 <-
  shoot.data4 %>%
  filter(is.na(growth) | !growth > 5 & !growth < -5) 
```

6. Remove NAs and rearrange columns
Also, re-factor categorical labels a bit for random effectsd

```{r}
shoot.data6 <- 
  shoot.data5 %>% 
  filter(!is.na(growth)) %>% 
  filter(!is.na(diameter)) %>% 
  mutate(interharvest_id = as.factor(interharvest_id)) %>% 
  separate(shoot_id, into = c("field", "harvest", "plant", "shoot"),
           remove = FALSE, sep = "-") %>% 
  mutate_at(vars(shoot_id, field, interharvest_id, plant, shoot), as.factor)

shootdf <- shoot.data6 %>%
  select(
    date,
    shoot_id,
    field,
    interharvest_id,
    plant,
    shoot,
    shoot_height,
    growth,
    diameter,
    day_post,
    mean_day,
    starts_with("temp"),
    starts_with("precip"),
    starts_with("Q"),
    starts_with("L")
  ) %>%
  ungroup()
```
```{r}
shootdf
skimr::skim(shootdf)
```

# Export Shoot Growth Data

```{r}
write_rds(shootdf, here("data", "cleaned", "shoots.rds"))
```
