==> wrangling/4-hopper-wrangling.Rmd <==
---
title: "Leafhopper data wrangling"
author: "Eric R. Scott"
date: "2019-11-5"
output: 
  html_notebook: 
    highlight: kate
    theme: yeti
    toc: yes
    toc_float: yes
    number_sections: yes
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, paged.print = FALSE)
```

*Last compiled: `r Sys.Date()`*

```{r packages, message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(lubridate)
library(conflicted)
conflict_prefer("filter", "dplyr")
conflict_prefer("lag", "dplyr")
conflict_prefer("here", "here")
```

# Purpose

Create a dataframe for analyzing leafhopper density as a function of weather.

# Read in data

```{r}
weather_lags <- read_rds(here("data", "cleaned", "weather_lags.rds"))
hoppers <- read_rds(here("data", "cleaned", "hoppers.rds"))
hourly.temps <- read_rds(here("data", "cleaned", "hourly-temps-hobo.rds"))
```

# Join in weather data

It looks like I started counting leafhoppers before I started measuring shoots.  There are NAs for day_post, harvest, and all the temperature data.  Blech.  Ok, really what I need to do is create a weather dataframe that has ALL the data, by date, with lagged temp, precip, interhavest, day post, mean_day so I can join it to both the plant and hopper datasets.

```{r}
hoppers2 <- left_join(hoppers, weather_lags, by = "date")
```

# Temp at time of counting
Find temperature at time of Count.Dtime (the time when leafhopper counts started).

```{r find temp at time of counting}
hourly <-
  hourly.temps %>%
  #create a column of times rounded to the nearest hour
  mutate(rounded = round_date(date_time, unit = "hour")) %>% 
  select(-date, -date_time)

hoppers3 <-
  hoppers2 %>% 
  mutate(rounded = round_date(date_time, unit = "hour"))

#join and clean up columns used only for joining
hoppers4 <-
  left_join(hoppers3, hourly, by = "rounded") %>% 
  select(-rounded) %>% 
  select(date_time, date, day, field, harvest, hoppers, n_leaves, temp_at_count = temp_c, everything())

hoppers4
```

# Calculate days post harvest

## First get harvest dates

Get harvest dates and create intervals

```{r}
harvest.dates <-
  hoppers4 %>% 
  group_by(date, field) %>% 
  summarise(Harvest = last(harvest)) %>% 
  filter(Harvest == "Y") %>% 
  ungroup() %>%
  select(-Harvest) %>% 
  add_row(date = max(hoppers4$date) + days(1), field = "B") %>% 
  group_by(field) %>%
  mutate(num = 1:n(),
         harvest_ID = paste0(field, num))
#on the last day of data collection field B was harvested

harvest.dates.A <- harvest.dates %>% filter(field == "A")
harvest.dates.B <- harvest.dates %>% filter(field == "B")
harvest.int.A <-
  (lag(harvest.dates.A$date) - days(1)) %--% (harvest.dates.A$date - days(1))
harvest.int.A <- harvest.int.A[-1]
# the last few days of data collection in field A
harvest.int.A[4] <- ymd("2017-07-22") %--% ymd("2017-07-24")

harvest.int.B <- (lag(harvest.dates.B$date) - days(1)) %--% (harvest.dates.B$date - days(1))

#the very first day of data collection was not immediately after a harvest.  Add that interval manually.

harvest.int.B[1] <- min(harvest.dates$date) %--% (harvest.dates.B$date[1] - days(1))
```

### Add interharvest column

Add a column for which interharvest the observation is in

```{r}
hoppers5 <-
  hoppers4 %>% 
  group_by(field) %>% 
  mutate(interharvest.no = case_when(date %within% harvest.int.A[1] & field == "A" ~ 1,
                                     date %within% harvest.int.A[2] & field == "A" ~ 2,
                                     date %within% harvest.int.A[3] & field == "A" ~ 3,
                                     date %within% harvest.int.A[4] & field == "A" ~ 4,
                                     date %within% harvest.int.B[1] & field == "B" ~ 1,
                                     date %within% harvest.int.B[2] & field == "B" ~ 2,
                                     date %within% harvest.int.B[3] & field == "B" ~ 3,
                                     TRUE                           ~ as.numeric(NA)) %>% as.factor())
```

### Calculate days post harvest

Also filter out the first interhavest period for field B because it's not immediately after a harvest

```{r}
#need to get interharvest periods.  Can't just borrow them from the plant data.
hoppers6 <-
  hoppers5 %>%
  group_by(field, interharvest.no) %>% 
  mutate(day_post = day - min(day, na.rm = TRUE),
         start = min(day, na.rm = TRUE),
         mean_day = mean(day, na.rm = TRUE)) %>%
  ungroup() %>% 
  arrange(field, date) %>% 
  filter(!(field == "B" & interharvest.no == 1)) 
```

### Final tidy up
```{r}
hoppers7 <- 
  hoppers6 %>% 
  #column type wrangling
  mutate(plant = paste0(field, plant)) %>% 
  mutate(interharvest_id = as.factor(paste0(field, interharvest.no)),
         counter = as.factor(counter),
         plant = as.factor(plant)) %>% 
  select(date_time, date, day, day_post, field, interharvest_id, plant, counter, hoppers, n_leaves, temp_at_count, starts_with("temp_"), starts_with("precip"), starts_with("Q"), starts_with("L"))
```


# Write to file

```{r save hopper data}
write_rds(hoppers7, here("data", "cleaned", "hopper-weather.rds"))
```
